+++ players/g2_player_old_2.py
